+++ R/distribution.R
@@ -1,4 +1,15 @@
 # Constructors for distributions
+evaluate_par <- function(par){
+  
+  if(is.symbol(par) | is.character(par)) {
+    par <- as.character(par)
+  } else{
+    par <- as.numeric(eval(par))
+  }
+  
+  par
+}
+
 
 #' Discrete gamma distribution
 #' 
@@ -9,8 +20,19 @@
 #' 
 #' @examples
 #' transitions <- list("S -> I" = d_gamma(rate = 1, shape = 5))
+#' transitions_dsl <- denim_dsl({S -> I = d_gamma(rate = 1, shape = 5)})
+#' # define model parameters as distributional parameters
+#' transitions_dsl <- denim_dsl({S -> I = d_gamma(rate = i_rate, shape = i_shape)})
 #' @export
 d_gamma <- function(rate, shape, dist_init = FALSE) {
+  # capture expression for evaluation
+  rate <- substitute(rate)
+  shape <- substitute(shape)
+  
+  # check whether it is a fixed value (numeric) or a model parameter (string or expression)
+  rate <- evaluate_par(rate)
+  shape <- evaluate_par(shape)
+  
   distr <- list(
     distribution = "gamma",
     rate = rate,
@@ -31,8 +53,17 @@ d_gamma <- function(rate, shape, dist_init = FALSE) {
 #' 
 #' @examples
 #' transitions <- list("I -> D" = d_weibull(0.6, 2))
+#' transitions <- denim_dsl({ I -> D = d_weibull(0.6, 2) })
 #' @export
 d_weibull <- function(scale, shape, dist_init = FALSE) {
+  # capture expression for evaluation
+  scale <- substitute(scale)
+  shape <- substitute(shape)
+  
+  # check whether it is a fixed value (numeric) or a model parameter (string or expression)
+  scale <- evaluate_par(scale)
+  shape <- evaluate_par(shape)
+  
   distr <- list(
     distribution = "weibull",
     scale = scale,
@@ -52,10 +83,14 @@ d_weibull <- function(scale, shape, dist_init = FALSE) {
 #' 
 #' @examples
 #' transitions <- list("I -> D" = d_exponential(0.3))
-#' 
-#'
+#' transitions <- denim_dsl({I -> D = d_exponential(0.3)})
 #' @export
 d_exponential <- function(rate, dist_init = FALSE) {
+  # capture expression for evaluation
+  rate <- substitute(rate)
+  # check whether it is a fixed value (numeric) or a model parameter (string or expression)
+  rate <- evaluate_par(rate)
+
   distr <- list(
     distribution = "exponential",
     rate = rate,
@@ -75,8 +110,17 @@ d_exponential <- function(rate, dist_init = FALSE) {
 #' 
 #' @examples
 #' transitions <- list("I -> D" = d_lognormal(3, 0.6))
+#' transitions <- denim_dsl({I -> D = d_lognormal(3, 0.6)})
 #' @export
 d_lognormal <- function(mu, sigma, dist_init = FALSE) {
+  # capture expression for evaluation
+  mu <- substitute(mu)
+  sigma <- substitute(sigma)
+  
+  # check whether it is a fixed value (numeric) or a model parameter (string or expression)
+  mu <- evaluate_par(mu)
+  sigma <- evaluate_par(sigma)
+  
   distr <- list(
     distribution = "lognormal",
     mu = mu,
@@ -97,7 +141,8 @@ d_lognormal <- function(mu, sigma, dist_init = FALSE) {
 #' @return a Distribution object for simulator
 #' 
 #' @examples
-#' transitions <- list("S->I"=mathexpr("beta*S/N"))
+#' transitions <- list("S->I"="beta*S/N")
+#' transitions <- denim_dsl({S->I=beta*S/N})
 #' # definition for parameters in the expression required
 #' params <- c(N = 1000, beta = 0.3)
 #' @export
@@ -136,17 +181,49 @@ constant <- function(x) {
 #' 
 #' Convert a vector of frequencies, percentages... into a distribution
 #' 
-#' @param ... a vector of values
+#' @param x a vector of values
 #' @return a Distribution object for simulator
 #' @param dist_init whether to distribute initial value across subcompartments following this distribution. (default to FALSE, meaning init value is always in the first compartment))
 #' 
 #' @examples
-#' transitions <- list("S->I"=nonparametric(0.1, 0.2, 0.5, 0.2))
+#' transitions <- list("S->I"=nonparametric( c(0.1, 0.2, 0.5, 0.2) ))
+#' transitions <- denim_dsl({S->I=nonparametric( c(0.1, 0.2, 0.5, 0.2) )})
+#' # you can also define a model parameter for the distribution
+#' transitions <- denim_dsl({S->I=nonparametric( dwelltime_dist )})
 #' @export
-nonparametric <- function(..., dist_init = FALSE) {
+nonparametric <- function(x, dist_init = FALSE) {
+  # x <- substitute(c(...))
+  # 
+  # # remove "c" 
+  # x <- as.character(x)[-1]
+  # 
+  # # try parsing ... 
+  # failed_parse <- FALSE
+  # try_eval <- sapply(parse(text = x), \(expr){
+  #   tryCatch(
+  #     eval(expr),
+  #     error = \(e) {
+  #       failed_parse <<- TRUE
+  #     }
+  #   )
+  # })
+  # x <- if(!failed_parse) {
+  #   try_eval
+  # } else{
+  #   x
+  # }
+  # 
+  # # if ... is a parameter and more than 1 parameter is given -> throw error
+  # if(all(is.character(x)) && length(x) > 1){
+  #   stop("nonparametric only takes 1 parameter as input")
+  # }
+  # 
+  
+  x <- substitute(x)
+  x <- evaluate_par(x)
   distr <- list(
     distribution = "nonparametric",
-    waitingTime = c(...),
+    waitingTime = x,
     dist_init = as.numeric(dist_init))
   
   class(distr) <- c("Distribution", class(distr))