+++ NEWS.md
@@ -1,3 +1,11 @@
+# denim 1.2.1
+* Minor runtime improvement
+* Add option to select color palette for plot
+
+# denim 1.2.0
+* Add denim DSL support
+* Add error handlers 
+
 # denim 1.1.0
 * Fix multinomial vs competing risks transition
 * Add distribute initial value feature